removed submit page and added functionality to end of upload step
