+++ src/controllers/uploadController.js
@@ -4,15 +4,44 @@ const upload = multer({ dest: 'uploads/' })
 
 const { Controller } = require('hmpo-form-wizard')
 
+const { readFile } = require('node:fs/promises')
+const { lookup } = require('mime-types')
+
+const apiRoute = 'http://127.0.0.1:8082/api/dataset/validate/file/request/'
+
 class UploadController extends Controller {
   middlewareSetup () {
     super.middlewareSetup()
     this.use('/upload', upload.single('datafile'))
   }
 
-  post (req, res, next) {
-    req.body.datafile = req.file
-    super.post(req, res, next)
+  async post (req, res, next) {
+    const formData = new FormData()
+    formData.append('dataset', req.sessionModel.get('dataset'))
+    formData.append('collection', req.sessionModel.get('data-subject'))
+    formData.append('organization', 'mockOrg')
+
+    const filePath = req.file.path
+    const file = new Blob([await readFile(filePath)], { type: lookup(filePath) })
+
+    formData.append('upload_file', file, req.file.originalname)
+
+    try {
+      // post the data to the api
+      const result = await fetch(apiRoute, {
+        method: 'POST',
+        body: formData
+      })
+
+      const json = await result.json()
+
+      console.log(json)
+
+      // send the response back to the user
+      res.send(json)
+    } catch (e) {
+      res.send(e)
+    }
   }
 }
